// .backup_fix_staff_in_reservation_20251221_120536/src/app/core/models/reservation.model.ts
export interface Reservation {
  id?: string;
  clientId: string;
  clientName?: string;
  date: string;
  startTime: string;
  endTime: string;
  
  assignedServerIds?: string[]; // Staff interne
  assignedTeamIds?: string[];   // Équipes externes (Multiple)
  
  selectedSlotId?: string;
  packId?: string;  
  notes?: string;
  status: 'CONFIRMED' | 'PENDING' | 'CANCELLED';
  totalPrice?: number;
  advance?: number;
  advanceOnly?: boolean;
  createdAt?: string;
}